2010-02-15 15:12 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/contrib/hbfbird/firebird.c     ! use native type for GC pointer item holder instead of casting       to 'void *'. Such casting may strip significant data when handler       is wider type then pointer.
  * harbour/contrib/hbnf/fttext.c
    ! fixed unnecessary casting

+++ harbour/contrib/hbnf/fttext.c
@@ -1839,7 +1839,7 @@ static int _ins_buff( int iLen )
       while( WriteLen > 0 )
       {
          /* position to beginning of write area */
-         if( hb_fsSeekLarge( handles[area], fpWrite, FS_SET ) != (unsigned long) fpWrite )
+         if( hb_fsSeekLarge( handles[area], fpWrite, FS_SET ) != fpWrite )
          {
             error[area] = hb_fsError();
             break;
